requirements.txt
python-telegram-bot==13.3
requests==2.25.1
flake8==3.8.3
pylint==2.5.3
